Added basic authorization

+++ routes/campgrounds.js
@@ -4,22 +4,9 @@ const catchAsync = require('../utils/catchAsync')
 const ExpressError = require('../utils/ExpressError')
 const Campground = require('../models/campground')
 const { campgroundSchema } = require('../schemas.js')
-const { isLoggedIn } = require('../middleware')
+const { isLoggedIn, isAuthor, validateCampground } = require('../middleware')
 
 
-const validateCampground = (req, res, next) => {
-    const { error } = campgroundSchema.validate(req.body)
-    if (error) {
-        // Joi delivers an array with possibly many error messages
-        // We need to join them
-        const msg = error.details.map(el => el.message).join(',')
-        console.log(msg)
-
-        throw new ExpressError(msg, 400)
-    } else {
-        next()
-    }
-}
 
 
 router.get('/', catchAsync(async (req, res) => {
@@ -33,13 +20,21 @@ router.get('/new', isLoggedIn, catchAsync(async (req, res) => {
 
 router.post('/', isLoggedIn, validateCampground, catchAsync(async (req, res, next) => {
     const campground = new Campground(req.body.campground)
+    campground.author = req.user._id
     await campground.save()
     req.flash('success', 'Successfully created a new campground!')
     res.redirect(`campgrounds/${campground._id}`)
 }))
 
 router.get('/:id', catchAsync(async (req, res) => {
-    const campground = await Campground.findById(req.params.id).populate('reviews')
+    // Nested populate for reviews
+    // The reviews shall have the author populated
+    const campground = await Campground.findById(req.params.id).populate({
+        path: 'reviews',
+        populate: {
+            path: 'author'
+        }
+    }).populate('author')
     if (!campground) {
         req.flash('error', 'Campground not found')
         return res.redirect('/campgrounds')
@@ -47,7 +42,7 @@ router.get('/:id', catchAsync(async (req, res) => {
     res.render('campgrounds/show', { campground })
 }))
 
-router.get('/:id/edit', isLoggedIn, catchAsync(async (req, res) => {
+router.get('/:id/edit', isLoggedIn, isAuthor, catchAsync(async (req, res) => {
     const campground = await Campground.findById(req.params.id)
     if (!campground) {
         req.flash('error', 'Campground not found')
@@ -59,7 +54,15 @@ router.get('/:id/edit', isLoggedIn, catchAsync(async (req, res) => {
 // Editing a campground
 router.put('/:id', isLoggedIn, validateCampground, catchAsync(async (req, res) => {
     const { id } = req.params
-    const campground = await Campground.findByIdAndUpdate(id, { ...req.body.campground })
+    // First we find the campground 
+    const campground = await Campground.findById(id)
+    // Then we check to see if the currently logged on user
+    // is the author of the campground
+    if (!campground.author.equals(req.user._id)) {
+        req.flash('error', 'You do not have permission to do that!')
+        return res.redirect(`/campgrounds/${id}`)
+    }
+    const campground2 = await Campground.findByIdAndUpdate(id, { ...req.body.campground })
     req.flash('success', 'Successfully updated campground!')
     res.redirect(`/campgrounds/${campground._id}`)
 }))
@@ -67,7 +70,15 @@ router.put('/:id', isLoggedIn, validateCampground, catchAsync(async (req, res) =
 // Deleting a campground
 router.delete('/:id', isLoggedIn, catchAsync(async (req, res) => {
     const { id } = req.params
-    const campground = await Campground.findByIdAndDelete(id)
+    // First we find the campground 
+    const campground = await Campground.findById(id)
+    // Then we check to see if the currently logged on user
+    // is the author of the campground
+    if (!campground.author.equals(req.user._id)) {
+        req.flash('error', 'You do not have permission to do that!')
+        return res.redirect(`/campgrounds/${id}`)
+    }
+    const campground2 = await Campground.findByIdAndDelete(id)
     req.flash('success', 'Successfully deleted campground!')
     res.redirect('/campgrounds')
 }))
